--- cardigan/stories/components/BannerCard/BannerCard.stories.mdx
@@ -1,30 +1,0 @@
-import BannerCard from '@weco/common/views/components/BannerCard/BannerCard';
-import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
-import * as stories from './BannerCard.stories.tsx';
-
-<Meta title={`Components/BannerCard`} component={BannerCard} />
-
-# BannerCard
-
-## Purpose
-
-A stand alone promo.
-
-### With a date range
-
-<Canvas>
-  <Story story={stories.withDateRange} />
-</Canvas>
-
-
-### Without a date range
-
-<Canvas>
-  <Story story={stories.withoutDateRange} />
-</Canvas>
-
-<ArgsTable />
-
-[Edit this on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/cardigan/stories/components/BannerCard/BannerCard.stories.mdx)
-
-
