+++ cardigan/stories/components/Cards/Cards.stories.mdx
@@ -1,0 +1,35 @@
+import { Meta, Canvas, Story, ArgsTable, Description } from '@storybook/addon-docs/blocks';
+import CompactCardReadme from '@weco/common/views/components/CompactCard/README.md';
+import BannerCardReadme from '@weco/common/views/components/BannerCard/README.md';
+import FeaturedCardReadme from '@weco/common/views/components/FeaturedCard/README.md';
+import * as stories from './Cards.stories.tsx';
+
+<Meta title={`Components/Cards`} />
+
+# Cards
+
+## `CompactCard`
+
+<Description>{CompactCardReadme}</Description>
+
+<Canvas>
+  <Story story={stories.compactCard} />
+</Canvas>
+
+## `BannerCard`
+
+<Description>{BannerCardReadme}</Description>
+
+<Canvas>
+  <Story story={stories.bannerCard} />
+</Canvas>
+
+
+## `FeaturedCard`
+
+<Description>{FeaturedCardReadme}</Description>
+
+<Canvas>
+  <Story story={stories.featuredCard} />
+</Canvas>
+
