+++ cardigan/stories/components/NewsletterPromo/NewsletterPromo.stories.mdx
@@ -1,0 +1,16 @@
+import { Meta, Canvas, Story, ArgsTable, Description } from '@storybook/addon-docs/blocks';
+import NewsletterPromo from '@weco/common/views/components/NewsletterPromo/NewsletterPromo';
+import Readme from '@weco/common/views/components/NewsletterPromo/README.md';
+import * as stories from './NewsletterPromo.stories.tsx';
+
+<Meta title={`Components/NewsletterPromo`} component={NewsletterPromo} />
+
+# NewsletterPromo
+
+<Description>{Readme}</Description>
+
+<Canvas>
+  <Story story={stories.basic} />
+</Canvas>
+
+<ArgsTable />
