+++ cardigan/stories/components/NewsletterPromo/NewsletterPromo.stories.tsx
@@ -1,0 +1,6 @@
+import NewsletterPromo from '@weco/common/views/components/NewsletterPromo/NewsletterPromo';
+
+const Template = args => <NewsletterPromo {...args} />;
+export const basic = Template.bind({});
+basic.args = {};
+basic.storyName = 'NewsletterPromo';
